migrate changes to deepseekv3 and remove unused config
